requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
contourpy==1.0.7
cycler==0.11.0
filelock==3.9.0
fonttools==4.38.0
huggingface-hub==0.11.1
idna==3.4
joblib==1.2.0
kiwisolver==1.4.4
llvmlite==0.39.1
matplotlib==3.6.3
numba==0.56.4
numpy==1.23.5
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pynndescent==0.5.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-adapt==0.0.82
pytorch-ignite==0.4.9
pytorch-metric-learning==1.7.1
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
scikit-learn==1.2.0
scipy==1.10.0
seaborn==0.12.2
six==1.16.0
threadpoolctl==3.1.0
timm==0.6.12
torch==1.13.1
torchmetrics==0.9.3
torchvision==0.14.1
tqdm==4.64.1
typing_extensions==4.4.0
umap-learn==0.5.3
urllib3==1.26.14
